+++ packages/faucet/src/api/httperror.ts
